requirements.txt
langchain == 0.1.12
sentence-transformers == 2.5.1
google-generativeai == 0.4.1
python-dotenv
pinecone-client
django == 5.0.3
pypdf
langchain-community == 0.0.28
langchain-google-genai == 0.0.11
flask
-e .

